Optionally register an SslContextFactory object.
Allow the optional registration of an SslContextFactory object with
the Environment. The specific use-case I have in mind is in
authenticating client-cert connections. At the Jersey/resource layer,
only the principal is accessible, so authentication of the certificate
must happen within the SSLContext. An SslContextFactory has to be
registered before the creation of a SSL-enabled Connector for it to be
used in an SSLContext.

An alternative approach is to create an X509TrustManager, which more
accurately resembles a client-cert authenticator. An inheriting
SslContextFactory would be created that returns this trust manager when
SslContextFactory#getTrustManagers() is called.

I prefer this approach as it seems more generally useful. It seems like
an object must be registered, rather than derived from a class or
configuration, particularly if authentication is decided by a DAO.